Create code - 2 For ode for a complex rocket four servo control onboard computer in Arduino:
In this code, we are using an altitude sensor to control the angle of four servos that are attached to the rocket. The altitude sensor is connected to the Arduino on pin A0, and the servos are connected to pins 9, 10, 11, and 12.

In the setup() function, we attach the servos to the control pins and set the altitude sensor pin as an input. In the loop() function, we read the altitude sensor value, map it to servo angles, and set the servos to those angles. We then wait 10 milliseconds before checking the altitude again.

Note that we are using the altitude sensor value to control the angle of the first and third servos in one direction (0-180 and 0-180, respectively), and the angle of the second and fourth servos in the opposite direction (180-0 and -180-0, respectively). This allows us